+++ web/upload/otherExpenseAccount/upload.txt
@@ -1,0 +1,1 @@
+This file is only intended to allow the creation of the tree on Git, thank you not to touch it.
